+++ src/euphorie/client/tests/test_functional_client.py
@@ -41,7 +41,6 @@ class SurveyTests(EuphorieFunctionalTestCase):
         # Start the survey
         browser.getControl(name="form.button.submit").click()
         session_url = browser.url.replace("/@@identification", "")
-        browser.getLink("Start Risk Identification").click()
         # Identify the risk
         browser.open(
             "%s/1/1/@@identification" % session_url
@@ -88,7 +87,6 @@ class SurveyTests(EuphorieFunctionalTestCase):
         # Start the survey
         browser.getControl(name="form.button.submit").click()
         session_url = browser.url.replace("/@@identification", "")
-        browser.getLink("Start Risk Identification").click()
         # Identify the top-5 risk
         browser.open(
             "%s/1/1/@@identification" % session_url
@@ -135,7 +133,6 @@ class SurveyTests(EuphorieFunctionalTestCase):
         # Start the survey
         browser.getControl(name="form.button.submit").click()
         session_url = browser.url.replace("/@@identification", "")
-        browser.getLink("Start Risk Identification").click()
         # Identify the top-5 risk
         browser.open(
             "%s/1/1/@@identification" % session_url